waveform/renderers/glwaveformrendererfilteredsignal: Add casts to GLfloat

+++ src/waveform/renderers/glwaveformrendererfilteredsignal.cpp
@@ -45,14 +45,19 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
         return;
     }
 
-    double firstVisualIndex = m_waveformRenderer->getFirstDisplayedPosition() * dataSize;
-    double lastVisualIndex = m_waveformRenderer->getLastDisplayedPosition() * dataSize;
-    const double lineWidth = (1.0 / m_waveformRenderer->getVisualSamplePerPixel()) + 1.0;
+    GLfloat firstVisualIndex = static_cast<GLfloat>(
+            m_waveformRenderer->getFirstDisplayedPosition() * dataSize);
+    GLfloat lastVisualIndex = static_cast<GLfloat>(
+            m_waveformRenderer->getLastDisplayedPosition() * dataSize);
+    const GLfloat lineWidth =
+            static_cast<GLfloat>(
+                    1.0 / m_waveformRenderer->getVisualSamplePerPixel()) +
+            1;
 
-    const int firstIndex = int(firstVisualIndex+0.5);
+    const int firstIndex = static_cast<int>(firstVisualIndex + 0.5);
     firstVisualIndex = firstIndex - firstIndex%2;
 
-    const int lastIndex = int(lastVisualIndex+0.5);
+    const int lastIndex = static_cast<int>(lastVisualIndex + 0.5);
     lastVisualIndex = lastIndex + lastIndex%2;
 
     // Reset device for native painting
@@ -86,8 +91,10 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
 
         //draw reference line
         glBegin(GL_LINES); {
-            glColor4f(m_axesColor_r, m_axesColor_g,
-                      m_axesColor_b, m_axesColor_a);
+            glColor4f(static_cast<GLfloat>(m_axesColor_r),
+                    static_cast<GLfloat>(m_axesColor_g),
+                    static_cast<GLfloat>(m_axesColor_b),
+                    static_cast<GLfloat>(m_axesColor_a));
             glVertex2f(firstVisualIndex,0);
             glVertex2f(lastVisualIndex,0);
         }
@@ -101,7 +108,10 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
             int firstIndex = math_max(static_cast<int>(firstVisualIndex), 0);
             int lastIndex = math_min(static_cast<int>(lastVisualIndex), dataSize);
 
-            glColor4f(m_lowColor_r, m_lowColor_g, m_lowColor_b, 0.8);
+            glColor4f(static_cast<GLfloat>(m_lowColor_r),
+                    static_cast<GLfloat>(m_lowColor_g),
+                    static_cast<GLfloat>(m_lowColor_b),
+                    0.8f);
             for (int visualIndex = firstIndex;
                     visualIndex < lastIndex;
                     visualIndex += 2) {
@@ -113,7 +123,10 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
                 glVertex2f(visualIndex,-1.f*lowGain*maxLow1);
             }
 
-            glColor4f(m_midColor_r, m_midColor_g, m_midColor_b, 0.85);
+            glColor4f(static_cast<GLfloat>(m_midColor_r),
+                    static_cast<GLfloat>(m_midColor_g),
+                    static_cast<GLfloat>(m_midColor_b),
+                    0.85f);
             for (int visualIndex = firstIndex;
                     visualIndex < lastIndex;
                     visualIndex += 2) {
@@ -125,7 +138,10 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
                 glVertex2f(visualIndex,-1.f * midGain * maxMid1);
             }
 
-            glColor4f(m_highColor_r, m_highColor_g, m_highColor_b, 0.9);
+            glColor4f(static_cast<GLfloat>(m_highColor_r),
+                    static_cast<GLfloat>(m_highColor_g),
+                    static_cast<GLfloat>(m_highColor_b),
+                    0.9f);
             for (int visualIndex = firstIndex;
                     visualIndex < lastIndex;
                     visualIndex += 2) {
@@ -165,7 +181,10 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
             int firstIndex = math_max(static_cast<int>(firstVisualIndex), 0);
             int lastIndex = math_min(static_cast<int>(lastVisualIndex), dataSize);
 
-            glColor4f(m_lowColor_r, m_lowColor_g, m_lowColor_b, 0.8);
+            glColor4f(static_cast<GLfloat>(m_lowColor_r),
+                    static_cast<GLfloat>(m_lowColor_g),
+                    static_cast<GLfloat>(m_lowColor_b),
+                    0.8f);
             for (int visualIndex = firstIndex;
                     visualIndex < lastIndex;
                     visualIndex += 2) {
@@ -178,7 +197,10 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
                 glVertex2f(visualIndex, lowGain * maxLow);
             }
 
-            glColor4f(m_midColor_r, m_midColor_g, m_midColor_b, 0.85);
+            glColor4f(static_cast<GLfloat>(m_midColor_r),
+                    static_cast<GLfloat>(m_midColor_g),
+                    static_cast<GLfloat>(m_midColor_b),
+                    0.85f);
             for (int visualIndex = firstIndex;
                     visualIndex < lastIndex;
                     visualIndex += 2) {
@@ -191,7 +213,10 @@ void GLWaveformRendererFilteredSignal::draw(QPainter* painter, QPaintEvent* /*ev
                 glVertex2f(visualIndex, midGain * maxMid);
             }
 
-            glColor4f(m_highColor_r, m_highColor_g, m_highColor_b, 0.9);
+            glColor4f(static_cast<GLfloat>(m_highColor_r),
+                    static_cast<GLfloat>(m_highColor_g),
+                    static_cast<GLfloat>(m_highColor_b),
+                    0.9f);
             for (int visualIndex = firstIndex;
                     visualIndex < lastIndex;
                     visualIndex += 2) {
